fix(profile): bind CQL LIMIT as i32 in list_by_account (was i64 → always errored)
ListProfilesByAccount cast the page limit to i64 before binding it into the
`… LIMIT ?` query. CQL `LIMIT` is a native `int`, so the driver serialized the
i64 as `BigInt` and every call failed type-checking before executing:

  bad CQL query: ... failed to serialize column [limit]:
  Failed to type check Rust type i64 against CQL type Native(Int): expected [BigInt]

Keep the clamped limit as i32 so it serializes as `int`. Add an integration
scenario that drives the repository method end-to-end (one profile, plus the
zero-row case that still binds LIMIT) — nothing exercised it before, which is
how the bug shipped. Both tests pass against live Scylla.

Fixes #623

+++ crates/services/profile/src/infrastructure/persistence/scylla_profile_repository.rs
@@ -332,7 +332,9 @@ impl ProfileRepository for ScyllaProfileRepository {
             created_at:   CqlTimestamp,
         }
 
-        let limit = limit.clamp(1, 100) as i64;
+        // CQL `LIMIT` is a native `int` (i32); binding an i64 makes the driver
+        // serialize it as BigInt and the query fails type-checking. Keep it i32.
+        let limit = limit.clamp(1, 100);
 
         let token = page_token
             .map(|t| -> Result<PageToken, ProfileError> {
